+++ innopoints/views/__init__.py
@@ -1,5 +1,6 @@
 from .authentication import *
 from .project import *
 from .activity import *
+from .file import *
 
 from .views import *  # temporary, until all views have been moved
