+++ innopoints/views/views.py
@@ -1,10 +1,6 @@
 """Application views"""
 
-import mimetypes
-
-import requests
-import werkzeug
-from flask import abort, jsonify, request, current_app
+from flask import abort, request
 from flask.views import MethodView
 from flask_login import login_required, current_user
 from marshmallow import ValidationError
@@ -13,13 +9,11 @@ from sqlalchemy.exc import IntegrityError
 
 from innopoints.extensions import db
 from innopoints.blueprints import api
-import innopoints.file_manager_s3 as file_manager
 from innopoints.models import (
     Color,
     Competence,
     Product,
     Size,
-    StaticFile,
     StockChange,
     StockChangeStatus,
     Variety,
@@ -32,8 +26,6 @@ from innopoints.schemas import (
     VarietySchema,
 )
 
-
-ALLOWED_MIMETYPES = {'image/jpeg', 'image/png', 'image/webp'}
 NO_PAYLOAD = ('', 204)
 
 
@@ -377,60 +369,6 @@ competence_api = CompetenceAPI.as_view('competence_api')
 api.add_url_rule('/competences/<int:compt_id>',
                  view_func=competence_api,
                  methods=('PATCH', 'DELETE'))
-
-
-# ----- Static File -----
-
-def get_mimetype(file: werkzeug.FileStorage) -> str:  # pylint: disable=no-member
-    """Return a MIME type of a Flask file object"""
-    if file.mimetype:
-        return file.mimetype
-
-    return mimetypes.guess_type(file.filename)[0]
-
-
-@api.route('/file/<namespace>', methods=['POST'])
-@login_required
-def upload_file(namespace):
-    """Upload a file to a given namespace."""
-    if 'file' not in request.files:
-        abort(400, {'message': 'No file attached.'})
-
-    file = request.files['file']
-
-    if not file.filename:
-        abort(400, {'message': 'The file doesn\'t have a name.'})
-
-    mimetype = get_mimetype(file)
-    if mimetype not in ALLOWED_MIMETYPES:
-        abort(400, {'message': f'Mimetype "{mimetype}" is not allowed'})
-
-    new_file = StaticFile(mimetype=mimetype, namespace=namespace)
-    db.session.add(new_file)
-    db.session.commit()
-    try:
-        file_manager.store(file.stream, str(new_file.id), new_file.namespace)
-    except requests.exceptions.HTTPError as exc:
-        print(exc)  # TODO: replace with proper logging
-        db.session.delete(new_file)
-        db.session.commit()
-        abort(400, {'message': 'Upload failed.'})
-    return jsonify(id=new_file.id, url=f'/file/{new_file.id}')
-
-
-@api.route('/file/<int:file_id>')
-def retrieve_file(file_id):
-    """Get the chosen static file"""
-    file = StaticFile.query.get_or_404(file_id)
-    url = f'{file.namespace}/{file.id}'
-    file_data = file_manager.retrieve(url)
-
-    if file_data is None:
-        abort(404)
-
-    response = current_app.make_response(file_data)
-    response.headers.set('Content-Type', file.mimetype)
-    return response
 
 
 # ----- Color -----
